Disjoin patterns without dependencies from prior partial matches. #224
When a query/pattern does not depend on preceding matches, it should be separated into an independent sub-network, such that it's only evaluated based on it's inputs chaning, and not based on the preceding partial matches.
The corolary to this is that a binding expression that does not depend on any preceding matches will only evaluate once (hence fixing the BindingEvaluationExceptionTest to introduce the dependency on the preceding fact match).